Fix changes based on review
